Using filters to collect cart endpoint data using StoreAPI
