+++ client/entrypoints/express-checkout/index.js
@@ -10,8 +10,8 @@ import {
 	getExpressCheckoutButtonAppearance,
 	getExpressCheckoutButtonStyleSettings,
 	getExpressCheckoutData,
-	isManualPaymentMethodCreation,
 	getPaymentMethodTypesForExpressMethod,
+	isManualPaymentMethodCreation,
 	normalizeLineItems,
 } from 'wcstripe/express-checkout/utils';
 import {
@@ -27,6 +27,7 @@ import {
 import { getStripeServerData } from 'wcstripe/stripe-utils';
 import { getAddToCartVariationParams } from 'wcstripe/utils';
 import 'wcstripe/express-checkout/compatibility/wc-order-attribution';
+import 'wcstripe/express-checkout/compatibility/wc-product-page';
 import './styles.scss';
 import {
 	EXPRESS_PAYMENT_METHOD_SETTING_AMAZON_PAY,
@@ -514,54 +515,20 @@ jQuery( function ( $ ) {
 		 * @return {Promise} Promise for the request to the server.
 		 */
 		addToCart: () => {
-			let productId = $( '.single_add_to_cart_button' ).val();
-
-			// Check if product is a variable product.
-			if ( $( '.single_variation_wrap' ).length ) {
-				productId = $( '.single_variation_wrap' )
-					.find( 'input[name="product_id"]' )
-					.val();
-			}
-
-			if ( $( '.wc-bookings-booking-form' ).length ) {
-				productId = $( '.wc-booking-product-id' ).val();
-			}
-
 			const data = {
 				qty: $( quantityInputSelector ).val(),
 			};
 
+			// Legacy support for variations.
 			if ( useLegacyCartEndpoints ) {
-				data.product_id = productId;
+				data.product_id = $( '.single_add_to_cart_button' ).val();
 				data.attributes = wcStripeECE.getAttributes().data;
-			} else {
-				data.id = productId;
-				data.variation = [];
-			}
 
-			// Add extension data to the POST body
-			const formData = $( 'form.cart' ).serializeArray();
-			$.each( formData, ( i, field ) => {
-				if ( /^(addon-|wc_)/.test( field.name ) ) {
-					if ( /\[\]$/.test( field.name ) ) {
-						const fieldName = field.name.substring(
-							0,
-							field.name.length - 2
-						);
-						if ( data[ fieldName ] ) {
-							data[ fieldName ].push( field.value );
-						} else {
-							data[ fieldName ] = [ field.value ];
-						}
-					} else {
-						data[ field.name ] = field.value;
-					}
-				}
-			} );
-
-			if ( useLegacyCartEndpoints ) {
 				return api.expressCheckoutAddToCartLegacy( data );
 			}
+
+			// BlocksAPI partial support (lacking support for variations).
+			data.variation = [];
 
 			return api.expressCheckoutAddToCart( data );
 		},
