+++ client/express-checkout/compatibility/wc-order-attribution.js
@@ -1,6 +1,3 @@
-/**
- * External dependencies
- */
 import { addFilter } from '@wordpress/hooks';
 import { extractOrderAttributionData } from 'wcstripe/blocks/utils';
